pyproject.toml
[project]
name = "unskild-vla-experiment"
version = "0.1.0"
description = "Vision-Language-Action (VLA) Training & Finetuning Framework"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.0",
    "torchvision>=0.15",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "huggingface-hub>=0.35.3",
    "lerobot>=0.4.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1",
]
transformers = [
    "transformers>=4.30",
    "accelerate>=0.20",
]
diffusers = [
    "diffusers>=0.20",
]
lerobot = [
    "lerobot>=0.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project.scripts]
vla-train = "src.cli.train:main"
vla-eval = "src.cli.eval:main"
vla-export = "src.cli.export:main"
